Fix DPR for swarm of quippers and sahuagin
The swam of quippers has two damages listed, one which does and
average of 14 damage and another which does an average of 7. My guess
is these were added, when it is one OR the other.

For a sahuagin, I am not sure where the DRP of 17 came from. They get
two attacks for a spear (4 or 5 average damage) and a bite for 3. This
is a total of 7 or 8 (depending on if the spear is wielded in 2 hands)

What is the source of these DPRs? Are there others which a far off
like this?